feat: create module for console operations
